split examples into one main per file
- this will be needed in a future CI test that uses the spec json
  validator against the output of these examples
